+++ packages/foam-vscode/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "foam-vscode-gollum" extension will be documented in this file.
 
+## [1.2.7] - 2025-08-15
+
+- Fix section heading completion for the existing page.
+- Fix wikilinks with no target and only a section heading.
+
 ## [1.2.6] - 2025-08-12
 
 - Move the cursor 1 character it ends on a new line.